requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
yfinance>=0.2.0
streamlit>=1.28.0
matplotlib>=3.5.0
plotly>=5.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.28.0
bcrypt>=4.0.0
PyJWT>=2.8.0
